# Faster boot for full ubuntu22 image, better volume sizing.