--- a/Makefile
+++ b/Makefile
@@ -7,9 +7,12 @@ check:
 	if ! git diff --exit-code :^Makefile &>/dev/null ; then echo "You have pending changes. Commit them first" ; exit 1 ; fi
 
 bump: check
-	sed -i 's|Default: "v1.*|Default: "$(VERSION)"|' cloudformation/template.yaml
+	sed -i 's|runs-on:v.*|runs-on:$(VERSION)|' cloudformation/template.yaml
+	cp cloudformation/template.yaml cloudformation/template-$(VERSION).yaml
 	sed -i 's|"version": "v1.*|"version": "$(VERSION)",|' package.json
-	if ! git diff --exit-code cloudformation/template.yaml ; then git commit -m "Bump template to $(VERSION)" Makefile package.json cloudformation/template.yaml && git tag -m "$(VERSION)" "$(VERSION)" ; fi
+
+commit:
+	if ! git diff --exit-code cloudformation/template.yaml ; then git commit -m "Bump template to $(VERSION)" Makefile package.json cloudformation/template.yaml cloudformation/template-$(VERSION).yaml && git tag -m "$(VERSION)" "$(VERSION)" ; fi
 
 login:
 	aws ecr-public get-login-password --region us-east-1 | docker login --username AWS --password-stdin public.ecr.aws/c5h5o9k1
@@ -23,8 +26,9 @@ push:
 
 s3-upload:
 	aws s3 cp ./cloudformation/template.yaml s3://runs-on/cloudformation/
+	aws s3 cp ./cloudformation/template-v*.yaml s3://runs-on/cloudformation/
 
-release: bump login build push s3-upload
+release: bump commit login build push s3-upload
 	docker tag public.ecr.aws/c5h5o9k1/runs-on/runs-on:$(VERSION) public.ecr.aws/c5h5o9k1/runs-on/runs-on:$(MAJOR_VERSION)
 	docker push public.ecr.aws/c5h5o9k1/runs-on/runs-on:$(MAJOR_VERSION)