--- a/bin/md-specs
+++ b/bin/md-specs
@@ -9,11 +9,11 @@ default_runner=$(node -e 'console.log(require("./src/constants").DEFAULT_RUNNER_
 echo "Default if no \`runner\` label provided: \`$default_runner\`."
 echo ""
 node -e 'console.log(JSON.stringify(require("./src/utils").objToArray(require("./src/constants").RUNNERS)))' | jq -r '
-  ["runner", "cpu", "family", "$/min (spot)", "$/min (on-demand)", "$/min (github)", "GitHub vs RunsOn"],
+  ["runner", "cpu", "family", "$/min (spot)", "$/min (on-demand)", "$/min (github)", "RunsOn vs GitHub"],
   ["---", "---", "---", "---", "---", "---", "---"],
   (.[] | [
     (.key | "`" + . + "`"), .cpu, (.family | join(", ")), .spot_price_per_min, .on_demand_price_per_min,
-    .github_price_per_min, (.github_ratio | select(. != null) | . + "x more expensive")
+    .github_price_per_min, (.github_ratio | (select(. != null) | . + "x cheaper") // "-" )
   ])
   | @tsv | gsub("\t"; " | ") | "| " + . + " |"
 '
@@ -25,8 +25,8 @@ default_image=$(node -e 'console.log(require("./src/constants").DEFAULT_IMAGE_SP
 echo "Default if no \`image\` label provided: \`$default_image\`."
 echo ""
 node -e 'console.log(JSON.stringify(require("./src/utils").objToArray(require("./src/constants").IMAGES)))' | jq -r '
-  ["image", "platform", "arc", "owner", "user", "name"],
-  ["---", "---", "---", "---", "---", "---"],
-  (.[] | [(.key | "`" + . + "`"), .platform, .arch, .owner, .user, .name])
+  ["image", "platform", "arch", "user", "name"],
+  ["---", "---", "---", "---", "---"],
+  (.[] | [(.key | "`" + . + "`"), .platform, .arch, .owner, .name])
   | @tsv | gsub("\t"; " | ") | "| " + . + " |"
 '