--- a/data/user_data/linux.sh.hbs
+++ b/data/user_data/linux.sh.hbs
@@ -1,7 +1,6 @@
 #!/bin/bash
 set -e
 set -o pipefail
-set -x
 
 # Note: this template must be valid for all linux variants (ubuntu, centos, etc), and all supported architectures (x64, arm64, etc)
 
@@ -18,6 +17,38 @@ RUNS_ON_AGENT_WAIT_TIMEOUT=20m
 # https://github.com/actions/actions-runner-controller/blob/master/docs/adrs/2022-10-17-runner-image.md
 RUNS_ON_AGENT_USER="runner"
 RUNS_ON_AGENT_DIR="/home/$RUNS_ON_AGENT_USER"
+RUNS_ON_LAUNCHED_AT={{{ launchedAt }}}
+
+if [ "$(uname -i)" = "aarch64" ]; then
+  echo "Detected arm64 architecture"
+  RUNS_ON_AGENT_ARCH="arm64"
+else
+  echo "Detected x64 architecture"
+  RUNS_ON_AGENT_ARCH="x64"
+fi
+
+# make sure env vars are available to all users
+echo '. /etc/environment' >> /etc/profile
+
+_run_with_timestamp() {
+  "$@"
+  local status=$?
+  local end=$(date -u +"%Y-%m-%dT%H:%M:%S%z")
+  local diff=$(($(date -u -d "$end" +"%s") - $(date -u -d "$RUNS_ON_LAUNCHED_AT" +"%s")))
+  echo "[$end] ${@/_/} (+${diff})" >> /tmp/runs-on-timings.log
+  return $status
+}
+
+_started() {
+  echo "Started"
+}
+
+_ready() {
+  echo "Ready"
+}
+
+echo "[$RUNS_ON_LAUNCHED_AT] launched" >> /tmp/runs-on-timings.log
+_run_with_timestamp _started
 
 _setup_watchdogs() {
   echo "Installing watchdogs..."
@@ -27,17 +58,16 @@ _setup_watchdogs() {
   sleep $RUNS_ON_AGENT_FULL_TIMEOUT && \
     echo "Full timeout reached. Shutting down instance." && _the_end &
 }
+_run_with_timestamp _setup_watchdogs
 
-_setup_watchdogs
-
-# make sure env vars are available to all users
-echo '. /etc/environment' >> /etc/profile
-
-echo "Storing SSH keys..."
-mkdir -p /root/runs-on/
-RUNS_ON_SSH_KEYS_FILE=/root/runs-on/authorized_keys
-# TODO: store those in S3 bucket instead?
-curl -Ls https://github.com/{{ join sshGithubUsernames delimiter=',' wrap='{}' }}.keys > "$RUNS_ON_SSH_KEYS_FILE"
+_store_ssh_keys() {
+  echo "Storing SSH keys..."
+  mkdir -p /root/runs-on/
+  RUNS_ON_SSH_KEYS_FILE=/root/runs-on/authorized_keys
+  # TODO: store those in S3 bucket instead?
+  curl -Ls https://github.com/{{ join sshGithubUsernames delimiter=',' wrap='{}' }}.keys > "$RUNS_ON_SSH_KEYS_FILE"
+}
+_run_with_timestamp _store_ssh_keys
 
 _setup_runner_user() {
   echo "Setting up runner user..."
@@ -46,6 +76,7 @@ _setup_runner_user() {
   echo "%sudo   ALL=(ALL:ALL) NOPASSWD:ALL" > /etc/sudoers
   echo "Defaults env_keep += \"DEBIAN_FRONTEND\"" >> /etc/sudoers
 }
+id "$RUNS_ON_AGENT_USER" || _run_with_timestamp _setup_runner_user
 
 _setup_ssh() {
   local homeUser="$RUNS_ON_AGENT_USER"
@@ -57,33 +88,16 @@ _setup_ssh() {
   chown -R "$homeUser":"$homeUser" "$homeDir"/.ssh
   chmod 700 "$homeDir"/.ssh && chmod 600 "$homeDir"/.ssh/authorized_keys
 }
+_run_with_timestamp _setup_ssh || true
 
-id "$RUNS_ON_AGENT_USER" || _setup_runner_user
-_setup_ssh || true
-
-
-EC2_TOKEN=$(curl -X PUT "http://169.254.169.254/latest/api/token" -H "X-aws-ec2-metadata-token-ttl-seconds: 21600")
-EC2_INSTANCE_TYPE=$(curl -H "X-aws-ec2-metadata-token: $EC2_TOKEN" -v http://169.254.169.254/latest/meta-data/instance-type)
-EC2_INSTANCE_ID=$(curl -H "X-aws-ec2-metadata-token: $EC2_TOKEN" -v http://169.254.169.254/latest/meta-data/instance-id)
-EC2_INSTANCE_LIFE_CYCLE=$(curl -H "X-aws-ec2-metadata-token: $EC2_TOKEN" -v http://169.254.169.254/latest/meta-data/instance-life-cycle)
-
-echo "Removing useless stuff..."
-rm -rf /etc/cron.d/* /etc/cron.hourly/* /etc/cron.daily/* /etc/cron.monthly/* /etc/cron.weekly/*
-
-echo "Storing preinstall scripts..."
-mkdir -p /root/runs-on/preinstall
-{{#preinstallScripts}}
-echo "{{{.}}}" | base64 -d > /root/runs-on/preinstall/{{@index}}-script.sh
-chmod a+x /root/runs-on/preinstall/{{@index}}-script.sh
-{{/preinstallScripts}}
-
-if [ "$(uname -i)" = "aarch64" ]; then
-  echo "Detected arm64 architecture"
-  RUNS_ON_AGENT_ARCH="arm64"
-else
-  echo "Detected x64 architecture"
-  RUNS_ON_AGENT_ARCH="x64"
-fi
+_fetch_instance_details() {
+  EC2_TOKEN=$(curl -s -X PUT "http://169.254.169.254/latest/api/token" -H "X-aws-ec2-metadata-token-ttl-seconds: 21600")
+  EC2_INSTANCE_TYPE=$(curl -s -H "X-aws-ec2-metadata-token: $EC2_TOKEN" http://169.254.169.254/latest/meta-data/instance-type)
+  EC2_INSTANCE_ID=$(curl -s -H "X-aws-ec2-metadata-token: $EC2_TOKEN" http://169.254.169.254/latest/meta-data/instance-id)
+  EC2_INSTANCE_LIFE_CYCLE=$(curl -s -H "X-aws-ec2-metadata-token: $EC2_TOKEN" http://169.254.169.254/latest/meta-data/instance-life-cycle)
+  echo "Instance type: $EC2_INSTANCE_TYPE\nInstance ID: $EC2_INSTANCE_ID\nInstance lifecycle: $EC2_INSTANCE_LIFE_CYCLE" > /tmp/runs-on-instance-details.log
+}
+_run_with_timestamp _fetch_instance_details || true
 
 cat >> /etc/environment <<EOF
 RUNS_ON_RUNNER_NAME="{{{runnerName}}}"
@@ -99,51 +113,72 @@ set -o allexport
 source /etc/environment
 set +o allexport
 
-_setup_hostname() {
-  echo "Setting up hostname..."
-  hostnamectl set-hostname --pretty "$RUNS_ON_AGENT_USER@$RUNS_ON_RUNNER_IP[$RUNS_ON_RUNNER_NAME]"
-  # echo "127.0.0.1 localhost $runnerName" >> /etc/hosts
+_resize_disk() {
+  if df -hT | grep -q /dev/nvme0n1p1 ; then
+    echo "Detected NVMe disk. Attempting to increase disk size to maximum..."
+    growpart /dev/nvme0n1 1 && resize2fs /dev/nvme0n1p1
+  fi
 }
+_run_with_timestamp _resize_disk || true
 
 _setup_agent() {
-  echo "Installing agent..."
-  mkdir -p "$RUNS_ON_AGENT_DIR"
-  time curl -o "actions-runner-linux.tar.gz" \
-    -L "https://github.com/actions/runner/releases/download/v$RUNS_ON_AGENT_VERSION/actions-runner-linux-$RUNS_ON_AGENT_ARCH-$RUNS_ON_AGENT_VERSION.tar.gz"
-  time tar xzf "./actions-runner-linux.tar.gz" -C $RUNS_ON_AGENT_DIR
+  if ! test -f "$RUNS_ON_AGENT_DIR/run.sh" ; then
+    echo "Agent not found. Installing..."
+    mkdir -p "$RUNS_ON_AGENT_DIR"
+    time curl -o "actions-runner-linux.tar.gz" \
+      -L "https://github.com/actions/runner/releases/download/v$RUNS_ON_AGENT_VERSION/actions-runner-linux-$RUNS_ON_AGENT_ARCH-$RUNS_ON_AGENT_VERSION.tar.gz"
+    time tar xzf "./actions-runner-linux.tar.gz" -C $RUNS_ON_AGENT_DIR
+  else
+    echo "Agent found. Skipping install..."
+  fi
+
   chown -R $RUNS_ON_AGENT_USER:$RUNS_ON_AGENT_USER $RUNS_ON_AGENT_DIR
+  if grep -E '^docker:' /etc/group &>/dev/null ; then
+    echo "Adding $RUNS_ON_AGENT_USER to docker group..."
+    usermod -aG docker $RUNS_ON_AGENT_USER || true
+  fi
 }
+_run_with_timestamp _setup_agent
+
+_setup_preinstall() {
+  echo "Storing preinstall scripts..."
+  mkdir -p /root/runs-on/preinstall
+  {{#preinstallScripts}}
+  echo "{{{.}}}" | base64 -d > /root/runs-on/preinstall/{{@index}}-script.sh
+  chmod a+x /root/runs-on/preinstall/{{@index}}-script.sh
+  {{/preinstallScripts}}
+
+  find /root/runs-on/preinstall -type f -print0 -name "*.sh" | sort --zero-terminated | xargs -r bash
+}
+_run_with_timestamp _setup_preinstall || true
 
 _setup_infos() {
   echo "Storing infos..."
   mkdir -p "$RUNS_ON_AGENT_DIR"
+  local timings=$(cat /tmp/runs-on-timings.log)
+  local instance_details=$(cat /tmp/runs-on-instance-details.log)
+  local ami_details=""
+  if test -s /imagegeneration/imagedata.json ; then
+    ami_details="$(cat /imagegeneration/imagedata.json | sed 's/[][]//g'),"
+  fi
   cat > "$RUNS_ON_AGENT_DIR/.setup_info" <<EOF
 [
+  $ami_details
   {
-    "group": "SSH Access",
-    "detail": "$RUNS_ON_AGENT_USER@$RUNS_ON_RUNNER_IP"
+    "group": "Runner Instance",
+    "detail": "SSH: $RUNS_ON_AGENT_USER@$RUNS_ON_RUNNER_IP\nName: $RUNS_ON_RUNNER_NAME\nArchitecture: $RUNS_ON_AGENT_ARCH\n$instance_details"
   },
   {
-    "group": "Runner",
-    "detail": "Name: $RUNS_ON_RUNNER_NAME\nArchitecture: $RUNS_ON_AGENT_ARCH\nInstance type: $EC2_INSTANCE_TYPE\nInstance ID: $EC2_INSTANCE_ID\nInstance lifecycle: $EC2_INSTANCE_LIFE_CYCLE"
+    "group": "Timings",
+    "detail": "${timings}"
   }
 ]
 EOF
 }
 
-_setup_preinstall() {
-  find /root/runs-on/preinstall -type f -print0 -name "*.sh" | sort --zero-terminated | xargs -r bash
-}
+_run_with_timestamp _ready
 
-_setup_hostname || true
-_setup_agent
-_setup_infos
-_setup_preinstall || true
-
-if grep -E '^docker:' /etc/group &>/dev/null ; then
-  echo "Adding $RUNS_ON_AGENT_USER to docker group..."
-  usermod -aG docker $RUNS_ON_AGENT_USER || true
-fi
+_setup_infos || true
 
 echo "Launching agent..."
 su - $RUNS_ON_AGENT_USER -c "cd $RUNS_ON_AGENT_DIR  && ./run.sh --jitconfig $RUNS_ON_AGENT_JIT_CONFIG"