## Faster boot for large images

`ubuntu22-full-x64`:

* Before: between 140 to 160s from launch to runner ready
* After: between 40 to 60s from launch to runner ready

<img width="469" alt="image" src="https://github.com/runs-on/runs-on/assets/6114/d7a67845-e15a-48e2-9508-0ee8cd613d95">


## Proper disk resizing

Volume is now properly resized, if given `hdd` size is greater than AMI size.

## New timings section in logs

Very useful to compare boot times between runner types / images.

<img width="474" alt="image" src="https://github.com/runs-on/runs-on/assets/6114/badab297-c421-4b5d-8628-c21b0a50b610">

## New license types

* standard
* sponsorship